chore: refactor operator deployment functions
