+++ pkg/utils/ownership.go
@@ -16,22 +16,14 @@ limitations under the License.
 
 package utils
 
-import (
-	"context"
-	"fmt"
-	"github.com/cloudnative-pg/cloudnative-pg/pkg/management/log"
-	v1 "k8s.io/api/apps/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-)
+import v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 // SetAsOwnedBy sets the controlled object as owned by a certain other
 // controller object with his type information
-func SetAsOwnedBy(controlled *metav1.ObjectMeta, controller metav1.ObjectMeta, typeMeta metav1.TypeMeta) {
+func SetAsOwnedBy(controlled *v1.ObjectMeta, controller v1.ObjectMeta, typeMeta v1.TypeMeta) {
 	isController := true
 
-	controlled.SetOwnerReferences([]metav1.OwnerReference{
+	controlled.SetOwnerReferences([]v1.OwnerReference{
 		{
 			APIVersion: typeMeta.APIVersion,
 			Kind:       typeMeta.Kind,
@@ -40,98 +32,4 @@ func SetAsOwnedBy(controlled *metav1.ObjectMeta, controller metav1.ObjectMeta, t
 			Controller: &isController,
 		},
 	})
-}
-
-// SetAsOwnedByOperatorDeployment sets the controlled object as owned by the operator deployment.
-//
-// IMPORTANT: The controlled resource must reside in the same namespace as the operator as described by:
-// https://kubernetes.io/docs/concepts/overview/working-with-objects/owners-dependents/
-func SetAsOwnedByOperatorDeployment(ctx context.Context,
-	kubeClient client.Client,
-	controlled *metav1.ObjectMeta,
-	operatorLabelSelector string,
-) error {
-	deployment, err := GetOperatorDeployment(ctx, kubeClient, controlled.Namespace, operatorLabelSelector)
-	if err != nil {
-		return err
-	}
-
-	// The deployment typeMeta is empty (kubernetes bug), so we need to explicitly populate it.
-	typeMeta := metav1.TypeMeta{
-		Kind:       "Deployment",
-		APIVersion: "apps/v1",
-	}
-	SetAsOwnedBy(controlled, deployment.ObjectMeta, typeMeta)
-
-	return nil
-}
-
-// GetOperatorDeployment find the operator deployment using labels
-// and then return the deployment object, in case we can't find a deployment
-// or we find more than one, we just return an error.
-func GetOperatorDeployment(
-	ctx context.Context,
-	kubeClient client.Client,
-	namespace, operatorLabelSelector string,
-) (*v1.Deployment, error) {
-	labelMap, err := labels.ConvertSelectorToLabelsMap(operatorLabelSelector)
-	if err != nil {
-		return nil, err
-	}
-	deployment, err := FindOperatorDeploymentByFilter(ctx,
-		kubeClient,
-		namespace,
-		client.MatchingLabelsSelector{Selector: labelMap.AsSelector()})
-	if err != nil {
-		return nil, err
-	}
-	if deployment != nil {
-		return deployment, nil
-	}
-
-	deployment, err = FindOperatorDeploymentByFilter(ctx,
-		kubeClient,
-		namespace,
-		client.HasLabels{"operators.coreos.com/cloudnative-pg.openshift-operators="})
-	if err != nil {
-		return nil, err
-	}
-	if deployment != nil {
-		return deployment, nil
-	}
-
-	return nil, fmt.Errorf("no deployment detected")
-}
-
-// FindOperatorDeploymentByFilter search in a defined namespace
-// looking for a deployment with the defined filter
-func FindOperatorDeploymentByFilter(ctx context.Context,
-	kubeClient client.Client,
-	namespace string,
-	filter client.ListOption,
-) (*v1.Deployment, error) {
-	logger := log.FromContext(ctx)
-
-	deploymentList := &v1.DeploymentList{}
-	err := kubeClient.List(
-		ctx,
-		deploymentList,
-		client.InNamespace(namespace),
-		filter,
-	)
-	if err != nil {
-		return nil, err
-	}
-	switch {
-	case len(deploymentList.Items) == 1:
-		return &deploymentList.Items[0], nil
-	case len(deploymentList.Items) > 1:
-		err = fmt.Errorf("more than one operator deployment running")
-		logger.Error(err, "more than one operator deployment found with the filter", "filter", filter)
-		return nil, err
-	}
-
-	err = fmt.Errorf("no operator deployment found")
-	logger.Error(err, "no operator deployment found with the filter", "filter", filter)
-	return nil, err
 }
